Externalize UIAccessibility extension
