+++ Sources/Public/CalendarView/CalendarView.swift
@@ -322,9 +322,46 @@ public final class CalendarView: UIView {
 
     lazy var doubleLayoutPassSizingLabel = DoubleLayoutPassSizingLabel(provider: self)
 
-    // MARK: Fileprivate
+    var content: CalendarViewContent
 
-    fileprivate var content: CalendarViewContent
+    var scrollToItemContext: ScrollToItemContext? {
+        willSet {
+            scrollToItemDisplayLink?.invalidate()
+        }
+    }
+
+    var calendar: Calendar {
+        content.calendar
+    }
+
+    var scrollMetricsMutator: ScrollMetricsMutator {
+        let scrollAxis: ScrollAxis = switch content.monthsLayout {
+            case .vertical: .vertical
+            case .horizontal: .horizontal
+        }
+
+        let scrollMetricsMutator: ScrollMetricsMutator = if let previousScrollMetricsMutator = _scrollMetricsMutator {
+            if scrollAxis != previousScrollMetricsMutator.scrollAxis {
+                ScrollMetricsMutator(
+                    scrollMetricsProvider: scrollView,
+                    scrollAxis: scrollAxis
+                )
+            } else {
+                previousScrollMetricsMutator
+            }
+        } else {
+            ScrollMetricsMutator(
+                scrollMetricsProvider: scrollView,
+                scrollAxis: scrollAxis
+            )
+        }
+
+        _scrollMetricsMutator = scrollMetricsMutator
+
+        return scrollMetricsMutator
+    }
+
+    // MARK: Fileprivate
 
     fileprivate var previousPageIndex: Int?
 
@@ -356,16 +393,6 @@ public final class CalendarView: UIView {
         gestureRecognizer.delegate = gestureRecognizerDelegate
         return gestureRecognizer
     }()
-
-    fileprivate var scrollToItemContext: ScrollToItemContext? {
-        willSet {
-            scrollToItemDisplayLink?.invalidate()
-        }
-    }
-
-    fileprivate var calendar: Calendar {
-        content.calendar
-    }
 
     // This hack is needed to prevent the scroll view from over-scrolling far past the content. This
     // occurs in 2 scenarios:
@@ -466,7 +493,7 @@ public final class CalendarView: UIView {
         return false
     }()
 
-    private var initialItemViewWasFocused = false {
+    internal var initialItemViewWasFocused = false {
         didSet {
             guard initialItemViewWasFocused != oldValue else { return }
             setNeedsLayout()
@@ -487,33 +514,6 @@ public final class CalendarView: UIView {
         // It's unclear whether values between 0 and 1 are possible, otherwise `max(scale, 1)` would
         // suffice.
         return scale > 0 ? scale : 1
-    }
-
-    private var scrollMetricsMutator: ScrollMetricsMutator {
-        let scrollAxis: ScrollAxis = switch content.monthsLayout {
-        case .vertical: .vertical
-        case .horizontal: .horizontal
-        }
-
-        let scrollMetricsMutator: ScrollMetricsMutator = if let previousScrollMetricsMutator = _scrollMetricsMutator {
-            if scrollAxis != previousScrollMetricsMutator.scrollAxis {
-                ScrollMetricsMutator(
-                    scrollMetricsProvider: scrollView,
-                    scrollAxis: scrollAxis
-                )
-            } else {
-                previousScrollMetricsMutator
-            }
-        } else {
-            ScrollMetricsMutator(
-                scrollMetricsProvider: scrollView,
-                scrollAxis: scrollAxis
-            )
-        }
-
-        _scrollMetricsMutator = scrollMetricsMutator
-
-        return scrollMetricsMutator
     }
 
     private var visibleItemsProvider: VisibleItemsProvider {
@@ -987,134 +987,6 @@ extension CalendarView: WidthDependentIntrinsicContentHeightProviding {
     }
 }
 
-// MARK: UIAccessibility
-
-public extension CalendarView {
-    // MARK: Public
-
-    override var isAccessibilityElement: Bool {
-        get { false }
-        set {}
-    }
-
-    override func accessibilityScroll(_ direction: UIAccessibilityScrollDirection) -> Bool {
-        guard
-            let firstVisibleMonth = visibleMonthRange?.lowerBound,
-            let lastVisibleMonth = visibleMonthRange?.upperBound,
-            let firstVisibleMonthDate = calendar.date(from: firstVisibleMonth.components),
-            let lastVisibleMonthDate = calendar.date(from: lastVisibleMonth.components),
-            let numberOfVisibleMonths = calendar.dateComponents(
-                [.month],
-                from: firstVisibleMonthDate,
-                to: lastVisibleMonthDate
-            )
-            .month
-        else {
-            return false
-        }
-
-        let proposedTargetMonth: Month
-        let scrollPosition: CalendarViewScrollPosition
-        switch (direction, content.monthsLayout) {
-        case (.up, .vertical), (.right, .horizontal):
-            proposedTargetMonth = Month(
-                era: lastVisibleMonth.era,
-                year: lastVisibleMonth.year,
-                month: lastVisibleMonth.month - numberOfVisibleMonths,
-                isInGregorianCalendar: lastVisibleMonth.isInGregorianCalendar
-            )
-            scrollPosition = .lastFullyVisiblePosition
-
-        case (.down, .vertical), (.left, .horizontal):
-            proposedTargetMonth = Month(
-                era: firstVisibleMonth.era,
-                year: firstVisibleMonth.year,
-                month: firstVisibleMonth.month + numberOfVisibleMonths,
-                isInGregorianCalendar: firstVisibleMonth.isInGregorianCalendar
-            )
-            scrollPosition = .firstFullyVisiblePosition
-
-        default:
-            return false
-        }
-
-        let firstMonth = content.monthRange.lowerBound
-        let lastMonth = content.monthRange.upperBound
-        let targetMonth = max(firstMonth, min(lastMonth, proposedTargetMonth))
-        guard let targetMonthDate = calendar.date(from: targetMonth.components) else { return false }
-
-        scroll(toMonthContaining: targetMonthDate, scrollPosition: scrollPosition, animated: false)
-
-        let targetMonthItem = content.monthHeaderItemProvider(targetMonth)
-        let targetMonthView = targetMonthItem._makeView()
-        targetMonthItem._setContent(onViewOfSameType: targetMonthView)
-        let accessibilityScrollText = targetMonthView.accessibilityLabel
-        UIAccessibility.post(notification: .pageScrolled, argument: accessibilityScrollText)
-
-        // ensure that scrolling related callbacks are still fired when performing scrolling via accessibility
-        if let visibleDayRange {
-            didScroll?(visibleDayRange, false)
-            didEndDragging?(visibleDayRange, true)
-            didEndDecelerating?(visibleDayRange)
-        }
-
-        return true
-    }
-
-    // MARK: Private
-
-    @objc
-    private func accessibilityElementFocused(_ notification: NSNotification) {
-        guard
-            let element = notification.userInfo?[UIAccessibility.focusedElementUserInfoKey] as? UIResponder,
-            let itemView = element.nextItemView()
-        else {
-            return
-        }
-
-        initialItemViewWasFocused = true
-
-        // If the accessibility element is not fully in view, programmatically scroll it to be centered.
-        let isElementFullyVisible: Bool
-        let viewFrameInCalendarView = itemView.convert(itemView.bounds, to: self)
-        switch scrollMetricsMutator.scrollAxis {
-        case .vertical:
-            let verticalBounds = CGRect(
-                x: 0,
-                y: layoutMargins.top,
-                width: bounds.width,
-                height: bounds.height - layoutMargins.top - layoutMargins.bottom
-            )
-            isElementFullyVisible = verticalBounds.contains(viewFrameInCalendarView)
-        case .horizontal:
-            let horizontalBounds = CGRect(
-                x: layoutMargins.left,
-                y: 0,
-                width: bounds.width - layoutMargins.left - layoutMargins.right,
-                height: bounds.height
-            )
-            isElementFullyVisible = horizontalBounds.contains(viewFrameInCalendarView)
-        }
-
-        if
-            !isElementFullyVisible,
-            let itemType = itemView.itemType,
-            case let .layoutItemType(layoutItemType) = itemType
-        {
-            switch layoutItemType {
-            case let .monthHeader(month):
-                let dateInTargetMonth = calendar.firstDate(of: month)
-                scroll(toMonthContaining: dateInTargetMonth, scrollPosition: .centered, animated: false)
-            case let .day(day):
-                let dateInTargetDay = calendar.startDate(of: day)
-                scroll(toDayContaining: dateInTargetDay, scrollPosition: .centered, animated: false)
-            default:
-                break
-            }
-        }
-    }
-}
-
 // MARK: - Calendar View Scroll Extension
 
 public extension CalendarView {
@@ -1497,7 +1369,7 @@ private extension UIScrollView {
 
 // MARK: `UIResponder` Next `ItemView`
 
-private extension UIResponder {
+extension UIResponder {
     func nextItemView() -> ItemView? {
         self as? ItemView ?? next?.nextItemView()
     }
